mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "MCP Server for EV Solar Charging Station Contracts"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.11.3",
    "web3>=6.15.1",
    "python-dotenv>=1.0.0",
    "groq>=0.4.1",
    "uvicorn>=0.25.0",
    "httpx>=0.26.0",
    "fastapi>=0.104.0",
]
